Add files via upload
This is an implementation of an SPI(Serial Peripheral Interface) protocol using the Verilog HDL. SPI is a full-duplex, faster than UART, I2C communication protocols. The design is done for one master and two slaves. The use of 8-bit registers is done to understand the operation and store the data. The dtf signal[data transfer] works with four different modes and the explanation of it can be given as below. 

dtf = 2'b00 means no communication 
         2'b01 means data from slave to master
         2'b10 means data from master to slave
         2'b11 means full duplex communication